feat: 使用 cobra 框架构建命令行
