+++ binding/golang/go.mod
@@ -1,7 +1,13 @@
 module github.com/lionsoul2014/ip2region/binding/golang
 
-go 1.17
+go 1.19
 
-require github.com/mitchellh/go-homedir v1.1.0
+require (
+	github.com/mitchellh/go-homedir v1.1.0
+	github.com/spf13/cobra v1.7.0
+)
 
-require github.com/yookoala/realpath v1.0.0 // indirect
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
