+++ maker/golang/ReadMe.md
@@ -14,17 +14,21 @@ make
 # `xdb` 数据生成
 
 通过 `xdb_maker gen` 命令生成 ip2region.xdb 二进制文件:
-```
-➜  golang git:(v2.0_xdb) ✗ ./xdb_maker gen
-./xdb_maker gen [command options]
-options:
- --src string    source ip text file path
- --dst string    destination binary xdb file path
+```bash
+$ ./xdb_maker gen                                                    
+Usage:
+  xdb_maker gen [flags]
+
+Flags:
+  -d, --dst string            destination binary xdb file path (required)
+  -h, --help                  help for gen
+  -i, --index-policy string   generate index policy (vector|btree) (default "vector")
+  -s, --src string            source ip text file path (required)
 ```
 
 例如，使用默认的 data/ip.merge.txt 作为源数据，生成一个 ip2region.xdb 到当前目录：
 ```bash
-➜  golang git:(v2.0_xdb) ✗ ./xdb_maker gen --src=../../data/ip.merge.txt --dst=./ip2region.xdb
+$ ./xdb_maker gen --src=../../data/ip.merge.txt --dst=./ip2region.xdb
 # 会看到一堆输出，最终会看到类似如下输出表示运行结束
 ...
 2022/06/16 16:38:48 maker.go:317: write done, with 13804 data blocks and (683591, 720221) index blocks
@@ -35,54 +39,58 @@ options:
 # `xdb` 数据查询
 
 通过 `xdb_maker search` 命令来测试查询输入的 ip：
-```
-➜  golang git:(v2.0_xdb) ✗ ./xdb_maker search
-./xdb_maker search [command options]
-options:
- --db string    ip2region binary xdb file path
+```bash
+$ ./xdb_maker search
+Usage:
+  xdb_maker search [flags]
+
+Flags:
+  -d, --db string   ip2region binary xdb file path (required)
+  -h, --help        help for search
 ```
 
 例如，使用自带的 xdb 文件来运行查询测试：
 ```bash
-➜  golang git:(v2.0_xdb) ✗ ./xdb_maker search --db=../../data/ip2region.xdb
+$ ./xdb_maker search --db=../../data/ip2region.xdb
 ip2region xdb search test program, commands:
 loadIndex : load the vector index for search speedup.
 clearIndex: clear the vector index.
-quit      : exit the test program
+quit      : exit the test program.
 ip2region>> 103.192.227.215
-{region:中国|0|香港|0|0, iocount:8, took:87.151µs}
-ip2region>> loadIndex
-vector index cached
+{region:中国|0|香港|0|0, iocount:8, took:1.390702ms}
 ip2region>> 39.114.2.16
-{region:韩国|0|0|0|SK宽带, iocount:2, took:50.005µs}
+{region:韩国|0|0|0|SK宽带, iocount:3, took:1.337932ms}
 ip2region>> 120.24.130.96
-{region:中国|0|广东省|深圳市|阿里云, iocount:2, took:32.805µs}
+{region:中国|0|广东省|深圳市|阿里云, iocount:3, took:1.387141ms}
 ip2region>> 
 ```
 
 # `xdb` 数据编辑
 
 通过 `xdb_maker edit` 命令来编辑原始的　IP 数据：
-```
-➜  golang git:(fr_editor) ✗ ./xdb_maker edit
-./xdb_maker edit [command options]
-options:
- --src string    source ip text file path
+```bash
+$ ./xdb_maker edit
+Usage:
+  xdb_maker edit [flags]
+
+Flags:
+  -h, --help         help for edit
+  -s, --src string   source ip text file path (required)
 ```
 
 例如，使用编辑器打开 `./data/ip.merge.txt` 会看到如下的操作面板：
 ```bash
-➜  golang git:(fr_editor) ✗ ./xdb_maker edit --src=../../data/ip.merge.txt
+$ ./xdb_maker edit --src=../../data/ip.merge.txt
 init the editor from source @ `../../data/ip.merge.txt` ... 
-all segments loaded, length: 683591, elapsed: 479.73743ms
-command list: 
+all segments loaded, length: 683843, elapsed: 503.631945ms
+command list:
   put [segment]        : put the specifield $segment
   put_file [file]      : put all the segments from the specified $file
   list [offset] [size] : list the first $size segments start from $offset
   save                 : save all the changes to the destination source file
   quit                 : exit the program
   help                 : print this help menu
-editor>>
+editor>> 
 ```
 
 通过 `put` 命令修改指定 IP 段的定位信息，例如：
@@ -109,18 +117,22 @@ editor>>
 # bench 测试
 
 如果你自主生成了 `xdb` 文件，请确保运行如下的 `xdb_maker bench` 命令来确保生成的的 `xdb` 文件的正确性：
-```
-➜  golang git:(v2.0_xdb) ✗ ./xdb_maker bench
-./xdb_maker bench [command options]
-options:
- --db string            ip2region binary xdb file path
- --src string           source ip text file path
- --ignore-error bool    keep going if bench failed
+```bash
+$ ./xdb_maker bench
+Usage:
+  xdb_maker bench [flags]
+
+Flags:
+  -d, --db string      ip2region binary xdb file path (required)
+  -h, --help           help for bench
+  -i, --ignore-error   keep going if bench failed (default "false")
+  -s, --src string     source ip text file path (required)
+
 ```
 
 例如：使用 data/ip.merge.txt 源文件来 bench 测试 data/ip2region.xdb 这个 xdb 文件：
 ```bash
-➜  golang git:(v2.0_xdb) ✗ ./xdb_maker bench --db=../../data/ip2region.xdb --src=../../data/ip.merge.txt
+$ ./xdb_maker bench --db=../../data/ip2region.xdb --src=../../data/ip.merge.txt
 # 会看到一堆输出，看到类似如下的数据表示 bench 测试通过了，否则就会报错
 ...
 try to bench segment: `224.0.0.0|255.255.255.255|0|0|0|内网IP|内网IP`
