+++ maker/golang/go.mod
@@ -1,3 +1,10 @@
 module github.com/lionsoul2014/ip2region/maker/golang
 
-go 1.17
+go 1.19
+
+require github.com/spf13/cobra v1.7.0
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
